Version 1.1

+++ src/main/java/de/finnos/southparkdownloader/processes/FfmpegProcess.java
@@ -37,7 +37,9 @@ public class FfmpegProcess implements CancelableProcess {
 
     @Override
     public void cancel() {
-        process.destroyForcibly();
+        if (process.isAlive()) {
+            process.destroyForcibly();
+        }
     }
 
     @Override
